+++ src/main/java/br/com/danielschiavo/customer/service/address/AddressService.java
@@ -1,11 +1,16 @@
 package br.com.danielschiavo.customer.service.address;
 
 
+import br.com.danielschiavo.customer.dto.request.address.RegisterAddressRequest;
+import br.com.danielschiavo.customer.dto.request.address.UpdateAddressRequest;
+import br.com.danielschiavo.customer.dto.response.address.DetailAddressResponse;
+import br.com.danielschiavo.customer.dto.response.address.ShowAddressesResponse;
 import br.com.danielschiavo.customer.model.entity.Address;
 import br.com.danielschiavo.customer.repository.AddressRepository;
 import br.com.danielschiavo.customer.mapper.AddressMapper;
 import br.com.danielschiavo.shared.exception.ValidationException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Example;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -25,52 +30,61 @@ public class AddressService {
 	
 	@Transactional
 	public void deleteAddressById(Long addressId, Long customerId) {
+		Address probe = new Address();
+		probe.setId(addressId);
+		probe.setCustomerId(customerId);
+		Example<Address> example = Example.of(probe);
+		if (!repository.exists(example)) {
+			throw new ValidationException("Could not delete this address, contact an administrator");
+		}
 		repository.deleteByIdAndCustomerId(addressId, customerId);
 	}
 	
-	public List<Address> getAllAddressesByCustomerId(Long customerId) {
-		return repository.findAllByCustomerId(customerId)
+	public List<ShowAddressesResponse> getAllAddressesByCustomerId(Long customerId) {
+		List<Address> addresses = repository.findAllByCustomerId(customerId)
 				.orElseThrow(() -> new ValidationException("Customer doesn't have any registered Address"));
+
+		return mapper.toListShowAddresses(addresses);
 	}
 	
-	public Address getAddressByIdAndCustomerId(Long addressId, Long customerId) {
-		return repository.findByIdAndCustomerId(addressId, customerId)
+	public DetailAddressResponse getAddressByIdAndCustomerId(Long addressId, Long customerId) {
+		Address address = repository.findByIdAndCustomerId(addressId, customerId)
 				.orElseThrow(() -> new ValidationException("Customer doesn't have a Address with id: " + addressId));
+
+		return mapper.toDetailAddress(address);
 	}
 	
 	@Transactional
-	public Address registerAddress(Long customerId, Address address) {
-		address.setCustomerId(customerId);
+	public DetailAddressResponse registerAddress(Long customerId, RegisterAddressRequest request) {
+		Address registerAddress = mapper.toEntity(request);
 
-		List<Address> addresses = getAllAddressesByCustomerId(customerId);
+		registerAddress.setCustomerId(customerId);
 
-		if (address.getIsDefault()) {
-			addresses.stream().filter(end -> end.getIsDefault().equals(true)).forEach(end -> end.setIsDefault(false));
-		}
+		defineOtherAddressAsIsDefaultFalse(customerId);
 
-		addresses.add(address);
-		repository.saveAll(addresses);
-		return address;
+		return mapper.toDetailAddress(repository.save(registerAddress));
 	}
 	
 	@Transactional
-	public Address updateAddress(Long addressId, Long customerId, Address updatedAddress) {
-		List<Address> allAddresses = getAllAddressesByCustomerId(customerId);
-		
-		Address address = allAddresses.stream().filter(end -> end.getId().equals(addressId))
-				.findFirst().orElseThrow(() -> new ValidationException("Address can't be updated because Customer doesn't have any registered Address with id: " + addressId));
+	public DetailAddressResponse updateAddress(Long addressId, Long customerId, UpdateAddressRequest request) {
+		Address address = repository.findByIdAndCustomerId(addressId, customerId)
+				.orElseThrow(() -> new ValidationException("Could not update this address, contact an administrator"));
 
-		mapper.updateAddress(updatedAddress, address);
+		mapper.updateAddress(request, address);
 
-		if (updatedAddress.getIsDefault()) {
-			allAddresses.stream().filter(end -> end.getIsDefault().equals(true) && !end.getId().equals(addressId))
-					.forEach(end -> end.setIsDefault(false));
+		if (request.isDefault()) {
+			defineOtherAddressAsIsDefaultFalse(customerId);
 		}
 
-		repository.saveAll(allAddresses);
-		return address;
+		return mapper.toDetailAddress(repository.save(address));
 	}
-	
+
+	public void defineOtherAddressAsIsDefaultFalse(Long customerId) {
+		repository.findByCustomerIdAndIsDefault(customerId, true).ifPresent(address -> {
+			address.setIsDefault(false);
+			repository.save(address);
+		});
+	}
 	
 //	------------------------------
 //	------------------------------
